localBattleAction context
