+++ modules/next/helpers/index.ts
@@ -1,2 +1,3 @@
 export * from "./useAppActionsAndState";
 export * from "./useBattleNavActions";
+export * from "./useLocalBattleActions";
