+++ modules/ui/src/contexts/battleActions.ts
@@ -11,7 +11,6 @@ export type BattleNavActions = {
     mode?: BattleMode | undefined,
     replace?: boolean
   ) => void;
-  newLocalBattle: (code: string, mode?: BattleMode | undefined) => void;
 };
 
 export const fakeBattleNavActions: BattleNavActions = {
@@ -20,7 +19,6 @@ export const fakeBattleNavActions: BattleNavActions = {
   toBattleLobby: () => console.log("to battle lobby"),
   toBattleControls: () => console.log("to battle controls"),
   toSession: (id, mode) => console.log("to session", id, mode),
-  newLocalBattle: (code, mode) => console.log("new local battle", code, mode),
 };
 
 export const BattleNavContext = createContext(fakeBattleNavActions);
